refactor: 코드 리뷰 피드백 반영
