+++ features/main/calendar/molecules/index.ts
@@ -1,1 +1,1 @@
-export { Calendar } from './calendar';
+export * from './calendar';
